requirements.txt
ascon==0.0.9
pycryptodome==3.20.0
pandas==2.1.4
matplotlib==3.8.2
numpy==1.26.2
